Reorganize pipeline Python spawners by concern
separation.js now holds only voice-separation stages (runSeparation,
runClearerVoice). Noise-reduction alternatives (runRnnoise, runDtln) move
to noiseReduce.js alongside DeepFilterNet3. Cross-pipeline ML enhancement
stages (harmonic exciter, vocal saturation, dereverb, BWE, click remover,
resonance suppressor) move into a new enhancement.js. Shared spawn helpers
are extracted into spawnPython.js to eliminate the duplicate spawn-Python
boilerplate that had accumulated across three files.

resonanceSuppressor.js is deleted (merged into enhancement.js). runVoiceFixer
and its deprecated Python script are removed.
